requirements.txt
Django
djangorestframework
gunicorn
dj-database-url
psycopg2